Implement disk adding and undo functionality
Switch to using a mutable bounded stack of immutable catalog states,
with the oldest items in the stack being evicted on overflow. Add
the ability to add disks so that there is something to undo, and
reorganize table and tree model code for better UI behaviour. Add
observable values so that menu items can be enabled/disabled as soon
as the model state changes.